pyproject.toml
[project]
name = "onlineshopfastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.45.0",
]
